Model_Testing/Models/TCN/requirements.txt
pandas==2.0.1
numpy==1.25.1
scikit-learn==1.3.2
tensorflow==2.14.0
keras-tcn==3.5.0
